move overview functions and tests to rest api app
