+++ chord_metadata_service/phenopackets/tests/test_api.py
@@ -282,41 +282,3 @@ class CreateInterpretationTest(APITestCase):
         response = get_response('interpretation-list',
                                 self.interpretation)
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-
-class OverviewTest(APITestCase):
-
-    def setUp(self) -> None:
-        # create 2 phenopackets for 2 individuals; each individual has 1 biosample;
-        # one of phenopackets has 1 phenotypic feature and 1 disease
-        self.individual_1 = m.Individual.objects.create(**c.VALID_INDIVIDUAL_1)
-        self.individual_2 = m.Individual.objects.create(**c.VALID_INDIVIDUAL_2)
-        self.metadata_1 = m.MetaData.objects.create(**c.VALID_META_DATA_1)
-        self.metadata_2 = m.MetaData.objects.create(**c.VALID_META_DATA_2)
-        self.phenopacket_1 = m.Phenopacket.objects.create(
-            **c.valid_phenopacket(subject=self.individual_1, meta_data=self.metadata_1)
-        )
-        self.phenopacket_2 = m.Phenopacket.objects.create(
-            id='phenopacket:2', subject=self.individual_2, meta_data=self.metadata_2
-        )
-        self.disease = m.Disease.objects.create(**c.VALID_DISEASE_1)
-        self.procedure = m.Procedure.objects.create(**c.VALID_PROCEDURE_1)
-        self.biosample_1 = m.Biosample.objects.create(**c.valid_biosample_1(self.individual_1, self.procedure))
-        self.biosample_2 = m.Biosample.objects.create(**c.valid_biosample_2(self.individual_2, self.procedure))
-        self.phenotypic_feature = m.PhenotypicFeature.objects.create(
-            **c.valid_phenotypic_feature(self.biosample_1, self.phenopacket_1)
-        )
-        self.phenopacket_1.biosamples.set([self.biosample_1])
-        self.phenopacket_2.biosamples.set([self.biosample_2])
-        self.phenopacket_1.diseases.set([self.disease])
-
-    def test_overview(self):
-        response = self.client.get('/api/overview')
-        response_obj = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertIsInstance(response_obj, dict)
-        self.assertEqual(response_obj['phenopackets'], 2)
-        self.assertEqual(response_obj['data_type_specific']['individuals']['count'], 2)
-        self.assertEqual(response_obj['data_type_specific']['biosamples']['count'], 2)
-        self.assertEqual(response_obj['data_type_specific']['phenotypic_features']['count'], 1)
-        self.assertEqual(response_obj['data_type_specific']['diseases']['count'], 1)
